// server/src/types/movie.type.ts
export type MovieData = {
  title: string;
  subtitle?: string;
  directors: string[];
  writers: string[];
  producers: string[];
  studios: string[];
  release_date: string;
  ISBN_EAN_UPC?: string;
  ISBN?: string;
  format: string | string[];
  duration: number;
  category: string | string[];
  summary: string;
  keywords: string[];
  targeted_audience: string;
  original_language: string;
  series: boolean;
  awards: string[];
  actors: string[];
  budget: number;
  box_office: number;
};